+++ src/pages/AboutUsPage.tsx
@@ -1,7 +1,12 @@
 import AboutUs from '../components/AboutUs';
+import MainLayout from '../shared/ui/MainLayout';
 
 const AboutUsPage = () => {
-  return <AboutUs />;
+  return (
+    <MainLayout>
+      <AboutUs />
+    </MainLayout>
+  );
 };
 
 export default AboutUsPage;
